[Optimize] Simplify workflow
