+++ package.json
@@ -12,23 +12,23 @@
     "node": ">=18.0.0"
   },
   "scripts": {
-    "server": "http-server . -p 8002 -o playground/index.html",
-    "postinstall": "gulp build-debug-infos && npm run build-declaration && npm run bundle-adapter && npm run fetch-ral && npm run install:native-pack-tool",
-    "build": "gulp build",
-    "build:dev": "gulp build-h5-source",
-    "build:min": "gulp build-h5-minified",
-    "build-declaration": "npm run build-const && gulp build-declarations",
-    "checkout-engine-version": "node ./scripts/checkout-engine-version.js",
-    "verify-engine-version": "node ./scripts/verify-engine-version.js",
-    "fetch-ral": "node ./scripts/fetch-ral.js",
-    "bundle-adapter": "node ./scripts/bundle-adapter.js",
-    "build-const": "node ./scripts/build-const.js",
+    "postinstall": "npm run build:debug-infos && npm run build:declaration && npm run build:adapter && npm run fetch-ral && npm run build:native-pack-tool",
+    "build": "npm run build:min && npm run build:declaration",
+    "build:adapter": "node ./scripts/bundle-adapter.js",
+    "build:const": "node ./scripts/build-const.js",
+    "build:dev": "npm run build:debug-infos && node ./scripts/build-h5-source",
+    "build:debug-infos": "node ./scripts/build-debug-infos",
+    "build:declaration": "npm run build:const && node ./scripts/build-declarations",
+    "build:min": "npm run build:debug-infos && node ./scripts/build-h5-minified",
+    "build:native-pack-tool": "cd ./scripts/native-pack-tool && npm install && npm run build",
     "clear": "node ./scripts/clear-cache.js",
-    "clear-platform": "node ./native/utils/clear-platform.js && node ./scripts/clear-file.js",
+    "clear:platform": "node ./native/utils/clear-platform.js && node ./scripts/clear-file.js",
+    "fetch-ral": "node ./scripts/fetch-ral.js",
+    "generate:doc-info": "node ./scripts/generate:doc-info",
+    "test": "tsc --noEmit && jest",
+    "update:engine-version": "node ./scripts/checkout-engine-version.js",
     "update:native-external": "node ./native/utils/update-external.js",
-    "test": "gulp test",
-    "generate-doc-info": "node ./scripts/generate-doc-info",
-    "install:native-pack-tool": "cd ./scripts/native-pack-tool && npm install && npm run build"
+    "verify:engine-version": "node ./scripts/verify-engine-version.js"
   },
   "repository": {
     "type": "git",
@@ -50,9 +50,6 @@
     "babelify": "^10.0.0",
     "browserify": "^17.0.0",
     "chalk": "^2.4.1",
-    "child_process": "^1.0.2",
-    "commander": "^2.20.3",
-    "dat.gui": "^0.7.5",
     "del": "^6.0.0",
     "eslint": "^8.44.0",
     "eslint-config-airbnb-base": "^15.0.0",
@@ -62,27 +59,19 @@
     "ftp": "0.3.10",
     "gulp": "^4.0.2",
     "gulp-rename": "^2.0.0",
-    "gulp-sequence": "^1.0.0",
     "gulp-uglify": "^3.0.2",
-    "gulp-zip": "^5.0.2",
-    "http-server": "^0.12.3",
     "jest": "^28.0.2",
     "jest-environment-jsdom": "^28.0.2",
     "jest-extended": "^1.1.0",
     "jest-matcher-deep-close-to": "^2.0.1",
-    "jsdom": "^19.0.0",
     "node-fetch": "^2.6.1",
-    "path": "^0.12.7",
     "pngjs": "^6.0.0",
     "progress": "^2.0.3",
-    "spectorjs": "^0.9.15",
-    "stats.js": "^0.17.0",
     "tar": "^6.1.15",
     "tmp": "^0.2.1",
     "tree-kill": "^1.2.0",
     "ts-jest": "^28.0.8",
     "typescript": "^4.3.5",
-    "vconsole": "^3.3.0",
     "vinyl-buffer": "^1.0.1",
     "vinyl-source-stream": "^2.0.0",
     "which": "^2.0.2",
