+++ scripts/bundle-adapter.js
@@ -1,9 +1,9 @@
 const ps = require('path');
 const fs = require('fs');
 const gulp = require('gulp');
-const babelify = require("babelify");
+const babelify = require('babelify');
 const browserify = require('browserify');
-const source = require("vinyl-source-stream");
+const source = require('vinyl-source-stream');
 const buffer = require('vinyl-buffer');
 const chalk = require('chalk').default;
 const uglify = require('gulp-uglify');
